requirements.txt
polygon-api-client==1.1.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
matplotlib==3.7.2
seaborn==0.12.2
ta==0.10.2
python-dotenv==1.0.0
tqdm==4.65.0
